# platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BrambiLab API
  version: 0.5.0
  description: |
    Base contract for web v1 (docs/architecture/web-v1.md §10).
    WEB-001: operation endpoints and the shared error shape. WEB-002: owner-only
    GitHub OAuth and sessions. WEB-003: private content, translations and
    immutable revisions. WEB-004: media library and authorized delivery under /media.
    WEB-005: publication, durable scheduling and withdrawal per locale; minimal public reader.
    Each later task adds its routes here first.
    Unsafe methods require an Origin (or Referer) equal to PUBLIC_ORIGIN; otherwise
    403 origin_rejected.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
paths:
...
  /admin/contents:
    get:
      operationId: listContents
      summary: Paginated contents, newest first, with a summary per translation.
      tags: [content]
      security: [{ session: [] }]
      parameters:
        - { name: kind, in: query, schema: { $ref: "#/components/schemas/ContentKind" } }
        - { name: project_id, in: query, description: Logs of this project., schema: { type: string, format: uuid } }
        - { name: archived, in: query, schema: { type: boolean, default: false } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses:
        "200":
          description: Page of contents.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContentPage" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/Invalid" }
    post:
      operationId: createContent
      summary: Creates a content with its first (empty) translation. Kind and parent never change later.
      tags: [content]
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, locale]
              additionalProperties: false
              properties:
                kind: { $ref: "#/components/schemas/ContentKind" }
                project_id: { type: string, format: uuid, description: Required for log; forbidden otherwise. }
                locale: { $ref: "#/components/schemas/Locale" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Content" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Invalid" }
  /admin/contents/{id}:
    parameters: [{ $ref: "#/components/parameters/ContentID" }]
    get:
      operationId: getContent
      summary: Content by id with translation summaries.
      tags: [content]
      security: [{ session: [] }]
      responses:
        "200":
          description: Content with translation summaries.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Content" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/contents/{id}/archive:
    parameters: [{ $ref: "#/components/parameters/ContentID" }]
    post:
      operationId: archiveContent
      summary: Archives without deleting history. Rejected with 409 while any translation is published (withdraw is WEB-005).
      tags: [content]
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      responses:
        "200":
          description: Archived.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Content" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /admin/contents/{id}/unarchive:
    parameters: [{ $ref: "#/components/parameters/ContentID" }]
    post:
      operationId: unarchiveContent
      summary: Makes an archived content active again.
      tags: [content]
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      responses:
        "200":
          description: Active again.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Content" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/contents/{id}/translations:
    parameters: [{ $ref: "#/components/parameters/ContentID" }]
    post:
      operationId: createTranslation
      summary: Adds a locale, empty or as an explicit copy (kind=copy revision, pending translation). Never machine-translated.
      tags: [content]
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [locale]
              additionalProperties: false
              properties:
                locale: { $ref: "#/components/schemas/Locale" }
                copy_from_locale: { $ref: "#/components/schemas/Locale" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Translation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Invalid" }
  /admin/contents/{id}/translations/{locale}:
    parameters:
      - { $ref: "#/components/parameters/ContentID" }
      - { $ref: "#/components/parameters/LocalePath" }
    get:
      operationId: getTranslation
      summary: Translation with its latest saved revision (null before the first save).
      tags: [content]
      security: [{ session: [] }]
      responses:
        "200":
          description: Translation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Translation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/contents/{id}/translations/{locale}/revisions:
    parameters:
      - { $ref: "#/components/parameters/ContentID" }
      - { $ref: "#/components/parameters/LocalePath" }
    get:
      operationId: listRevisions
      summary: Revision history, newest first.
      tags: [content]
      security: [{ session: [] }]
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
      responses:
        "200":
          description: Page of revisions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RevisionPage" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: saveRevision
      summary: Saves a full immutable snapshot (manual or auto). Unchanged snapshot → 200 created=false.
      description: |
        expected_version must equal the translation's latest_version, else 409 version_conflict
        with current_version. Retrying with the same Idempotency-Key returns the same result.
        Body limits: 1 MiB request, depth 32, 20000 nodes (web-v1.md §7.1).
      tags: [content]
      security: [{ session: [] }]
      parameters:
        - { $ref: "#/components/parameters/CSRF" }
        - name: Idempotency-Key
          in: header
          description: Client-generated key (8–128 chars [A-Za-z0-9_-]) per save operation.
          schema: { type: string, pattern: "^[A-Za-z0-9_-]{8,128}$" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SaveRevisionRequest" }
      responses:
        "200":
          description: Snapshot identical to the latest revision; nothing created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaveRevisionResponse" }
        "201":
          description: Revision created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaveRevisionResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "422": { $ref: "#/components/responses/Invalid" }
  /admin/contents/{id}/translations/{locale}/revisions/{version}:
    parameters:
      - { $ref: "#/components/parameters/ContentID" }
      - { $ref: "#/components/parameters/LocalePath" }
      - { $ref: "#/components/parameters/Version" }
    get:
      operationId: getRevision
      summary: One saved revision. The private preview renders from this.
      tags: [content]
      security: [{ session: [] }]
      responses:
        "200":
          description: Revision.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Revision" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/contents/{id}/translations/{locale}/revisions/{version}/restore:
    parameters:
      - { $ref: "#/components/parameters/ContentID" }
      - { $ref: "#/components/parameters/LocalePath" }
      - { $ref: "#/components/parameters/Version" }
    post:
      operationId: restoreRevision
      summary: Creates a new revision (kind=restore) with the old snapshot. The original is never modified.
      tags: [content]
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_version]
              additionalProperties: false
              properties:
                expected_version: { type: integer, minimum: 0 }
      responses:
        "201":
          description: Restored as a new revision.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaveRevisionResponse" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  parameters:
    CSRF:
      name: X-CSRF-Token
      in: header
      required: true
      description: Session CSRF token from /auth/me. Origin must also equal PUBLIC_ORIGIN.
      schema: { type: string }
    ContentID:
      { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    LocalePath:
      { name: locale, in: path, required: true, schema: { $ref: "#/components/schemas/Locale" } }
    Version:
      { name: version, in: path, required: true, schema: { type: integer, minimum: 1 } }
    Page:
      { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize:
      { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-bl_session
      description: Opaque session cookie (HttpOnly, Secure, SameSite=Lax, Path=/). Named bl_session on http://localhost.
  responses:
    Forbidden:
      description: origin_rejected or csrf_failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Unknown resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: version_conflict (with current_version), editorial_conflict (with state), published_content, scheduled_content, has_published_logs, schedule_exists, no_active_schedule, not_published, locale_exists or slug_taken.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Invalid:
      description: validation_failed (per-field messages in fields), media_not_available, publish_blocked (one entry per blocking field or asset) or idempotency_key_reused.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TooLarge:
      description: Request body above 1 MiB.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthenticated:
      description: No live owner session.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    EditorialStatus: { type: string, enum: [unpublished, published, withdrawn] }
...
    EditorialState:
      type: object
      required: [status, editorial_version, latest_version, active_job]
      properties:
        status: { $ref: "#/components/schemas/EditorialStatus" }
        editorial_version: { type: integer, minimum: 0 }
        latest_version: { type: integer, description: Latest saved draft revision. }
        published_version: { type: [integer, "null"] }
        published_at: { type: [string, "null"], format: date-time }
        first_published_at: { type: [string, "null"], format: date-time }
        withdrawn_at: { type: [string, "null"], format: date-time }
        route: { type: [string, "null"], description: Current public API path while published. }
        active_job:
          oneOf:
            - $ref: "#/components/schemas/PublicationJob"
            - type: "null"
        time_zone: { type: string, const: America/Mexico_City }
...
    ContentKind: { type: string, enum: [project, article, log] }
    Locale: { type: string, enum: [es, en] }
    RevisionKind: { type: string, enum: [manual, auto, restore, copy] }
    TranslationSummary:
      type: object
      required: [locale, latest_version, published]
      properties:
        locale: { $ref: "#/components/schemas/Locale" }
        latest_version: { type: integer, minimum: 0 }
        title: { type: [string, "null"], description: Title of the latest revision. }
        published: { type: boolean }
        updated_at: { type: [string, "null"], format: date-time }
    Content:
      type: object
      required: [id, kind, created_at, translations]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/ContentKind" }
        project_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
        archived_at: { type: [string, "null"], format: date-time }
        translations: { type: array, items: { $ref: "#/components/schemas/TranslationSummary" } }
    ContentPage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Content" } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    Document:
      type: object
      description: Canonical document v1 (web-v1.md §7.1), validated strictly by the API.
      required: [type, content]
      properties:
        type: { const: doc }
        content: { type: array, items: { $ref: "#/components/schemas/Node" } }
...
    Snapshot:
      type: object
      required: [title, slug, body]
      additionalProperties: false
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$", maxLength: 120 }
        summary: { type: string, maxLength: 500 }
        body: { $ref: "#/components/schemas/Document" }
        seo:
          type: object
          additionalProperties: false
          properties:
            title: { type: string, maxLength: 70 }
            description: { type: string, maxLength: 160 }
        project_fields: { $ref: "#/components/schemas/ProjectFields" }
        category_id: { type: [string, "null"], format: uuid }
        cover_asset_id: { type: [string, "null"], format: uuid, description: Ready image used as cover (WEB-004). }
        tag_ids: { type: array, maxItems: 20, uniqueItems: true, items: { type: string, format: uuid } }
    Revision:
      allOf:
        - $ref: "#/components/schemas/Snapshot"
        - type: object
          required: [version, kind, created_at, body_schema_version]
          properties:
            version: { type: integer, minimum: 1 }
            kind: { $ref: "#/components/schemas/RevisionKind" }
            created_at: { type: string, format: date-time }
            body_schema_version: { type: integer, const: 1 }
            restored_from_version: { type: [integer, "null"] }
            copied_from_locale: { type: [string, "null"] }
    RevisionSummary:
      type: object
      required: [version, kind, created_at, title]
      properties:
        version: { type: integer }
        kind: { $ref: "#/components/schemas/RevisionKind" }
        created_at: { type: string, format: date-time }
        title: { type: string }
        restored_from_version: { type: [integer, "null"] }
    RevisionPage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/RevisionSummary" } }
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    Translation:
      type: object
      required: [content_id, locale, latest_version, published, latest]
      properties:
        content_id: { type: string, format: uuid }
        locale: { $ref: "#/components/schemas/Locale" }
        latest_version: { type: integer, minimum: 0 }
        published: { type: boolean }
        latest:
          oneOf:
            - $ref: "#/components/schemas/Revision"
            - type: "null"
    SaveRevisionRequest:
      type: object
      required: [expected_version, kind, snapshot]
      additionalProperties: false
      properties:
        expected_version: { type: integer, minimum: 0 }
        kind: { type: string, enum: [manual, auto] }
        snapshot: { $ref: "#/components/schemas/Snapshot" }
    SaveRevisionResponse:
      type: object
      required: [created, revision]
      properties:
        created: { type: boolean }
        revision: { $ref: "#/components/schemas/Revision" }
...
    Error:
      type: object
      description: Body of every non-2xx response except health probes.
      required: [code, message, request_id]
      properties:
        code:
          type: string
          examples: [not_found, internal, validation_failed]
        message:
          type: string
        fields:
          type: object
          description: Per-field validation messages.
          additionalProperties: { type: string }
        request_id:
          type: string
        current_version:
          type: integer
          description: On version_conflict (latest_version) and editorial_conflict (editorial_version).
        state:
          $ref: "#/components/schemas/EditorialState"
          description: Only on editorial_conflict, the current editorial state.
...